Update Makefile
Possibility of making both interface or only one (do <make epidec> or <make epicode> in interface folder). Also possible to make encode and decode without interface in the folder Encoding (do <make encode> or <make decode>) for the executables without interface the arguments weren't modified. For decoding use the arguments used in the segmentation test and for encoding use the usual one.